Extract Calculation.Value to its own file
