+++ packages/alfa-css/src/value/calculation/calculation.ts
@@ -1,8 +1,6 @@
 import { Hash } from "@siteimprove/alfa-hash";
-import { Mapper } from "@siteimprove/alfa-mapper";
 import { Option, None } from "@siteimprove/alfa-option";
 import { Parser } from "@siteimprove/alfa-parser";
-import { Selective } from "@siteimprove/alfa-selective";
 import { Slice } from "@siteimprove/alfa-slice";
 import { Result } from "@siteimprove/alfa-result";
 
@@ -11,7 +9,7 @@ import * as json from "@siteimprove/alfa-json";
 import { Token } from "../../syntax/token";
 import { Function } from "../../syntax/function";
 
-import { Value } from "../../value";
+import { Value as CSSValue } from "../../value";
 import {
   Angle,
   Dimension,
@@ -24,6 +22,7 @@ import { Unit } from "../unit";
 
 import { Expression } from "./expression";
 import { Kind } from "./kind";
+import { Value } from "./value";
 
 const { delimited, either, filter, flatMap, map, option, pair } = Parser;
 
@@ -40,7 +39,7 @@ const { isPercentage } = Percentage;
  */
 export class Calculation<
   out D extends Calculation.Dimension = Calculation.Dimension
-> extends Value<"calculation"> {
+> extends CSSValue<"calculation"> {
   public static of(expression: Expression): Calculation {
     return new Calculation(
       expression.reduce({
@@ -188,119 +187,6 @@ export namespace Calculation {
     | `${Numeric.Dimension}-percentage`
     | "number";
 
-  function angleResolver(angle: Angle): Angle<"deg"> {
-    return angle.withUnit("deg");
-  }
-
-  function lengthResolver<U extends Unit.Length = "px">(
-    resolver: Mapper<Length<Unit.Length.Relative>, Length<U>>
-  ): Mapper<Length, Length<"px"> | Length<U>> {
-    return (length) =>
-      length.isRelative() ? resolver(length) : length.withUnit("px");
-  }
-
-  export class Value extends Expression {
-    public static of(value: Numeric): Value {
-      return new Value(value);
-    }
-
-    private readonly _value: Numeric;
-
-    private constructor(value: Numeric) {
-      super();
-
-      this._value = value;
-    }
-
-    public get type(): "value" {
-      return "value";
-    }
-
-    public get kind(): Kind {
-      const value = this._value;
-
-      if (isPercentage(value)) {
-        return Kind.of("percentage");
-      }
-
-      if (isLength(value)) {
-        return Kind.of("length");
-      }
-
-      if (isAngle(value)) {
-        return Kind.of("angle");
-      }
-
-      return Kind.of();
-    }
-
-    public get value(): Numeric {
-      return this._value;
-    }
-
-    public reduce<L extends Unit.Length = "px", P extends Numeric = Numeric>(
-      resolver: Expression.Resolver<L, P>
-    ): Value {
-      return Value.of(
-        Selective.of(this._value)
-          .if(isLength, lengthResolver(resolver.length))
-          .if(isAngle, angleResolver)
-          .if(isPercentage, resolver.percentage)
-          .get()
-      );
-    }
-
-    public toLength(): Option<Length> {
-      if (isLength(this.value)) {
-        return Option.of(this.value);
-      }
-
-      return None;
-    }
-
-    public toNumber(): Option<Number> {
-      if (isNumber(this.value)) {
-        return Option.of(this.value);
-      }
-
-      return None;
-    }
-
-    public toPercentage(): Option<Percentage> {
-      if (isPercentage(this.value)) {
-        return Option.of(this.value);
-      }
-
-      return None;
-    }
-
-    public equals(value: unknown): value is this {
-      return value instanceof Value && value._value.equals(this._value);
-    }
-
-    public toJSON(): Value.JSON {
-      return {
-        type: "value",
-        value: this._value.toJSON(),
-      };
-    }
-
-    public toString(): string {
-      return `${this._value}`;
-    }
-  }
-
-  export namespace Value {
-    export interface JSON extends Expression.JSON {
-      type: "value";
-      value: Numeric.JSON;
-    }
-  }
-
-  export function isValueExpression(value: unknown): value is Value {
-    return value instanceof Value;
-  }
-
   /**
    * {@link https://drafts.csswg.org/css-values/#calculation-tree-operator-nodes}
    */
@@ -388,7 +274,7 @@ export namespace Calculation {
         operand.reduce(resolver)
       );
 
-      if (isValueExpression(fst) && isValueExpression(snd)) {
+      if (Value.isValueExpression(fst) && Value.isValueExpression(snd)) {
         if (isNumber(fst.value) && isNumber(snd.value)) {
           return Value.of(Number.of(fst.value.value + snd.value.value));
         }
@@ -448,7 +334,7 @@ export namespace Calculation {
         operand.reduce(resolver)
       );
 
-      if (isValueExpression(operand)) {
+      if (Value.isValueExpression(operand)) {
         const { value } = operand;
 
         if (isNumber(value)) {
@@ -508,7 +394,7 @@ export namespace Calculation {
         operand.reduce(resolver)
       );
 
-      if (isValueExpression(fst) && isValueExpression(snd)) {
+      if (Value.isValueExpression(fst) && Value.isValueExpression(snd)) {
         let multipler: number | undefined;
         let value!: Numeric;
 
@@ -573,7 +459,7 @@ export namespace Calculation {
         operand.reduce(resolver)
       );
 
-      if (isValueExpression(operand)) {
+      if (Value.isValueExpression(operand)) {
         const { value } = operand;
 
         if (isNumber(value)) {
